+++ ChangeLog.md
@@ -11,6 +11,18 @@ ChangeLog
 * #139: We now _always_ return `DateTimeImmutable` from any method. This could
   potentially have big implications if you manipulate Date objects anywhere.
 
+
+3.3.3 (2014-??-??)
+------------------
+
+* #142: `CANCEL` and `REPLY` messages now include the `DTSTART` from the
+  original event.
+* #143: `SCHEDULE-AGENT` on the `ORGANIZER` property is respected.
+* #144: `PARTSTAT=NEEDS-ACTION` is now set for new invites, if no `PARTSTAT` is
+  set to support the inbox feature of iOS.
+* #147: Bugs related to scheduling all-day events.
+
+
 3.3.2 (2014-09-19)
 ------------------